Rearchitect `ForEach` to compute the views eagerly
This simplifies the generic bounds quite significantly and since we are
recomputing this anyway, there is no real point in deferring it to event
dispatch/render time.

For lazily computed views, we may look into providing something like
SwiftUI's LazyVStack/LazyHStack in the future. This would likely require
a bit of rearchitecting, since we couldn't just prerender the tree but
would likely require calling back and forth across the FFI boundary so
the underlying LazyVStack/LazyHStack could actually render lazily.